assignment3/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
draggan==1.1.6
executing==2.1.0
face-alignment==1.4.1
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gradio==3.34.0
gradio_client==1.4.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.43.0
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.13.5
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.4.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.1
termcolor==2.5.0
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
websockets==12.0
yarl==1.17.1
